[P3] Prevent Unintended ME Migration Application
